Make Encounter Menu Work- not right now
Here's a ribbon fix, do NOT merge this till we have the link on the actual page working to the guides. Dead page is still dead. 
However, it's here for posterity and to help anyone who tries to mess with it later get up to speed on this issue faster.